refactor(COR-1811): Add generic tooltip function
