+++ packages/app/src/domain/variants/variants-stacked-bar-chart-tile.tsx
@@ -7,11 +7,10 @@ import { StackedBarTooltipData, StackedChart } from '~/components/stacked-chart'
 import { useBarConfig } from '~/domain/variants/logic/use-bar-config';
 import { InteractiveLegend } from '~/components/interactive-legend';
 import { useList } from '~/utils/use-list';
-import { TooltipData } from '~/components/time-series-chart/components';
-import { isDefined, isPresent } from 'ts-is-present';
 import { TooltipSeriesList } from '~/components/time-series-chart/components/tooltip/tooltip-series-list';
 import { space } from '~/style/theme';
 import { useCurrentDate } from '~/utils/current-date-context';
+import { reorderAndFilter } from '~/domain/variants/logic/reorder-and-filter';
 
 interface VariantsStackedBarChartTileProps {
   title: string;
@@ -25,40 +24,6 @@ interface VariantsStackedBarChartTileProps {
 }
 
 const alwaysEnabled: (keyof VariantChartValue)[] = [];
-
-/**
- * Check if the key metricProperty exists
- * @param config
- */
-const hasMetricProperty = (config: any): config is { metricProperty: string } => {
-  return 'metricProperty' in config;
-};
-
-/**
- * Only variants that have a greater occurrence than 0 must be shown in the tooltip, except when the user narrows down
- * the total amount of visible variants by selecting one or more from the legend
- * @param context - Tooltip data context
- * @param selectionOptions - Currently selected variants
- */
-const reorderAndFilter = (context: TooltipData<VariantChartValue & StackedBarTooltipData>, selectionOptions: StackedBarConfig<VariantChartValue>[]) => {
-  const metricAmount = context.config.length;
-  const totalMetricAmount = selectionOptions.length;
-  const hasSelectedMetrics = metricAmount !== totalMetricAmount; // Check whether the user has selected any variants from the interactive legend.
-
-  /* Filter out any variants that have an occcurrence value of 0 */
-  const filteredValues = Object.fromEntries(
-    Object.entries(context.value).filter(([key, value]) => (key.includes('occurrence') ? value !== 0 && isPresent(value) && !isNaN(Number(value)) : value))
-  ) as VariantChartValue;
-
-  /* Rebuild tooltip data context with filtered values */
-  const reorderContext = {
-    ...context,
-    config: [...context.config.filter((value) => !hasMetricProperty(value) || filteredValues[value.metricProperty] || hasSelectedMetrics)].filter(isDefined),
-    value: !hasSelectedMetrics ? filteredValues : context.value,
-  };
-
-  return reorderContext as TooltipData<VariantChartValue & StackedBarTooltipData>;
-};
 
 /**
  * Variant bar chart component
@@ -101,7 +66,9 @@ export const VariantsStackedBarChartTile = ({ title, description, helpText, tool
         config={barChartConfig}
         disableLegend
         timeframe={variantTimeFrame}
-        formatTooltip={(data) => <TooltipSeriesList data={reorderAndFilter(data, selectionOptions)} hasTwoColumns={hasTwoColumns} />}
+        formatTooltip={(data) => (
+          <TooltipSeriesList data={reorderAndFilter<StackedBarTooltipData, StackedBarConfig<VariantChartValue>>(data, selectionOptions)} hasTwoColumns={hasTwoColumns} />
+        )}
       />
     </ChartTile>
   );
